+++ COMPILE_RFASM_BIN/Compiler/CompilationException.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace RFASM_COMPILER.RFASM_BIN
+namespace RedFoxAssembly.Compiler
 {
     internal class CompilationException : Exception
     {
