+++ COMPILE_RFASM_BIN/Compiler/PreCompilationException.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace RFASM_COMPILER.RFASM_BIN
+namespace RedFoxAssembly.Compiler
 {
     internal class PreCompilationException : Exception
     {
