+++ COMPILE_RFASM_BIN/Compiler/Tokens/InstructionToken.cs
@@ -1,4 +1,5 @@
-﻿using RFASM_COMPILER.TOKEN_PARSER;
+﻿using RedFoxAssembly.Compiler;
+using RedFoxAssembly.TokenParser;
 using System;
 using System.Collections;
 using System.Collections.Generic;
@@ -6,7 +7,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace RFASM_COMPILER.RFASM_BIN.TOKENS
+namespace RedFoxAssembly.Compiler.Tokens
 {
     internal class InstructionToken: AbstractToken
     {
