+++ COMPILE_RFASM_BIN/Compiler/Tokens/RawValueToken.cs
@@ -1,16 +1,18 @@
-﻿using RFASM_COMPILER.TOKEN_PARSER;
+﻿using RedFoxAssembly.Compiler;
+using RedFoxAssembly.TokenParser;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace RFASM_COMPILER.RFASM_BIN.TOKENS
+namespace RedFoxAssembly.Compiler.Tokens
 {
-    internal class RegisterAddressToken: AbstractValueToken
+    internal class RawValueToken: AbstractValueToken
     {
-        public RegisterAddressToken(string value, RFASMCompilerMetadata meta): base(value, meta, AddressingMode.REGISTER)
+        public RawValueToken(string value, RFASMCompilerMetadata meta) : base(value, meta, AddressingMode.RAW)
         {
+
         }
 
         public override byte[] GetBytes()
@@ -18,14 +20,14 @@ namespace RFASM_COMPILER.RFASM_BIN.TOKENS
             byte[] ba = Convert.FromHexString(RawValue);
             if (ba.Length != meta.DataWidth)
             {
-                throw new CompilationException("Data width " + ba.Length + " of RegisterAddressToken " + RawValue + " does not match expected width " + meta.DataWidth);
+                throw new CompilationException("Data width " + ba.Length + " of RawValueToken " + RawValue + " does not match expected width " + meta.DataWidth);
             }
             return CompilerUtils.FitToDataWidth(meta.DataWidth, ba);
         }
 
         public override string ParseRawValue(string reallyRawValue)
         {
-            return reallyRawValue.Replace("*", "", StringComparison.Ordinal);
+            return reallyRawValue.Replace("#", "", StringComparison.Ordinal);
         }
 
         public override bool CheckFollowing(IToken[] following)
