+++ COMPILE_RFASM_BIN/RedFoxAssembly.csproj
